move code to utils/plot_map.py
move all map/cartopy related code into plot_map.py to simplify utils/plot.py

bugs fix
